[SFU] Rename sfu_utilities.py to roads.py; change to UNIX endings
